adds cld

+++ src/services/profiles/index.js
@@ -76,10 +76,12 @@ router.get("/me", authenticateToken, async (req, res, next) => {
   }
 });
 router.put("/:id/picture", authenticateToken, cloudMulter.single("image"), async (req, res, next) => {
-  console.log("req file",req.file.path)
-    try {
-      const uploadImage = await profileSchema.findByIdAndUpdate(req.params.id,{ image:req.file.path },{ runValidators: true, new: true });
-      res.status(201).send(uploadImage)
+  try {
+    console.log("**************** I M G ****************")
+		return res.json({ body: req.body, file: req.file })
+      //security code
+      //const uploadImage = await profileSchema.findByIdAndUpdate(req.params.id,{ image:req.file },{ runValidators: true, new: true });
+      //res.status(201).send(uploadImage)
     } catch (error) {
         next(error)
     }
@@ -107,9 +109,21 @@ router.get("/:id", authenticateToken, async (req, res, next) => {
 });
 router.put("/:id", authenticateToken, async (req, res, next) => {
   try {
+    const post = { ...req.body }
+		const author = await profileSchema.findById(req.params.id, {
+			_id: 0,
+			userName: 1,
+		})
+		if (author.userName !== req.user.name) {
+			const error = new Error(
+				`Please do not try to change profile with ${req.params.id}`
+			)
+			error.httpStatusCode = 403
+			return next(error)
+		}
     const profile = await profileSchema.findByIdAndUpdate(
       req.params.id,
-      req.body,
+      post,
       { runValidators: true, new: true }
     );
     if (profile) {
@@ -125,6 +139,17 @@ router.put("/:id", authenticateToken, async (req, res, next) => {
 });
 router.delete("/:id", authenticateToken, async (req, res, next) => {
 	try {
+		const author = await profileSchema.findById(req.params.id, {
+			_id: 0,
+			userName: 1,
+		})
+		if (author.userName !== req.user.name) {
+			const error = new Error(
+				`Please do not try to delete profile with ${req.params.id}`
+			)
+			error.httpStatusCode = 403
+			return next(error)
+		}
 		const profile = await profileSchema.findByIdAndDelete(req.params.id)
 		if (profile) {
 			res.send("deleted")
@@ -137,9 +162,4 @@ router.delete("/:id", authenticateToken, async (req, res, next) => {
 		next(error)
 	}
 })
-
-
-
-
-
 module.exports = router;
